use builder api when creating semantic tokens
This mostly solves the performance issue. Opening zi/src/editor.rs used
to take ~80ms building the tree. Now it's nearer 1ms.

The reason doing repeated inserts was really slow was the summary
additions and subtractions. crop does a tonne of them and they're not
cheap since they contain bitmaps. This would take over 95% of the time
of building the tree.